Revert "feat: Sync from noir (#11796)"
This reverts commit 3840e8e01c12656229651baaae120858b1c99911.
