+++ noir/noir-repo/.github/workflows/formatting.yml
@@ -123,7 +123,18 @@ jobs:
         uses: actions/checkout@v4
       
       - name: Download nargo binary
-        uses: ./.github/actions/download-nargo
+        uses: actions/download-artifact@v4
+        with:
+          name: nargo
+          path: ./nargo
+            
+      - name: Set nargo on PATH
+        run: |
+          nargo_binary="${{ github.workspace }}/nargo/nargo"
+          chmod +x $nargo_binary
+          echo "$(dirname $nargo_binary)" >> $GITHUB_PATH
+          export PATH="$PATH:$(dirname $nargo_binary)"
+          nargo -V
 
       - name: Format stdlib
         working-directory: ./noir_stdlib
