+++ noir/noir-repo/compiler/noirc_evaluator/src/acir/mod.rs
@@ -762,12 +762,7 @@ impl<'a> Context<'a> {
                     None
                 };
 
-                self.acir_context.assert_neq_var(
-                    lhs,
-                    rhs,
-                    self.current_side_effects_enabled_var,
-                    assert_payload,
-                )?;
+                self.acir_context.assert_neq_var(lhs, rhs, assert_payload)?;
             }
             Instruction::Cast(value_id, _) => {
                 let acir_var = self.convert_numeric_value(*value_id, dfg)?;
