+++ noir/noir-repo/compiler/noirc_evaluator/src/brillig/brillig_gen/brillig_globals.rs
@@ -313,7 +313,6 @@ mod tests {
                     2,
                     "Expected just a `Return`, but got more than a single opcode"
                 );
-                // TODO: Bring this back (https://github.com/noir-lang/noir/issues/7306)
                 // assert!(matches!(&artifact.byte_code[0], Opcode::Return));
             } else if func_id.to_u32() == 2 {
                 assert_eq!(
@@ -431,7 +430,6 @@ mod tests {
                     30,
                     "Expected enough opcodes to initialize the globals"
                 );
-                // TODO: Bring this back (https://github.com/noir-lang/noir/issues/7306)
                 // let Opcode::Const { destination, bit_size, value } = &artifact.byte_code[0] else {
                 //     panic!("First opcode is expected to be `Const`");
                 // };
