+++ noir/noir-repo/compiler/noirc_frontend/src/ast/mod.rs
@@ -253,6 +253,12 @@ pub enum UnresolvedTypeExpression {
     AsTraitPath(Box<AsTraitPath>),
 }
 
+impl Recoverable for UnresolvedType {
+    fn error(span: Span) -> Self {
+        UnresolvedType { typ: UnresolvedTypeData::Error, span }
+    }
+}
+
 impl std::fmt::Display for GenericTypeArg {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
