+++ noir/noir-repo/compiler/noirc_frontend/src/ast/visitor.rs
@@ -4,7 +4,7 @@ use noirc_errors::Span;
 use crate::{
     ast::{
         ArrayLiteral, AsTraitPath, AssignStatement, BlockExpression, CallExpression,
-        CastExpression, ConstrainExpression, ConstructorExpression, Expression, ExpressionKind,
+        CastExpression, ConstrainStatement, ConstructorExpression, Expression, ExpressionKind,
         ForLoopStatement, ForRange, Ident, IfExpression, IndexExpression, InfixExpression, LValue,
         Lambda, LetStatement, Literal, MemberAccessExpression, MethodCallExpression,
         ModuleDeclaration, NoirFunction, NoirStruct, NoirTrait, NoirTraitImpl, NoirTypeAlias, Path,
@@ -294,7 +294,7 @@ pub trait Visitor {
         true
     }
 
-    fn visit_constrain_statement(&mut self, _: &ConstrainExpression) -> bool {
+    fn visit_constrain_statement(&mut self, _: &ConstrainStatement) -> bool {
         true
     }
 
@@ -855,9 +855,6 @@ impl Expression {
             ExpressionKind::MethodCall(method_call_expression) => {
                 method_call_expression.accept(self.span, visitor);
             }
-            ExpressionKind::Constrain(constrain) => {
-                constrain.accept(visitor);
-            }
             ExpressionKind::Constructor(constructor_expression) => {
                 constructor_expression.accept(self.span, visitor);
             }
@@ -1151,6 +1148,9 @@ impl Statement {
             StatementKind::Let(let_statement) => {
                 let_statement.accept(visitor);
             }
+            StatementKind::Constrain(constrain_statement) => {
+                constrain_statement.accept(visitor);
+            }
             StatementKind::Expression(expression) => {
                 expression.accept(visitor);
             }
@@ -1199,7 +1199,7 @@ impl LetStatement {
     }
 }
 
-impl ConstrainExpression {
+impl ConstrainStatement {
     pub fn accept(&self, visitor: &mut impl Visitor) {
         if visitor.visit_constrain_statement(self) {
             self.accept_children(visitor);
