+++ noir/noir-repo/compiler/noirc_frontend/src/hir/comptime/display.rs
@@ -6,7 +6,7 @@ use noirc_errors::Span;
 use crate::{
     ast::{
         ArrayLiteral, AsTraitPath, AssignStatement, BlockExpression, CallExpression,
-        CastExpression, ConstrainExpression, ConstructorExpression, Expression, ExpressionKind,
+        CastExpression, ConstrainStatement, ConstructorExpression, Expression, ExpressionKind,
         ForBounds, ForLoopStatement, ForRange, GenericTypeArgs, IfExpression, IndexExpression,
         InfixExpression, LValue, Lambda, LetStatement, Literal, MatchExpression,
         MemberAccessExpression, MethodCallExpression, Pattern, PrefixExpression, Statement,
@@ -573,12 +573,6 @@ fn remove_interned_in_expression_kind(
                 ..*call
             }))
         }
-        ExpressionKind::Constrain(constrain) => ExpressionKind::Constrain(ConstrainExpression {
-            arguments: vecmap(constrain.arguments, |expr| {
-                remove_interned_in_expression(interner, expr)
-            }),
-            ..constrain
-        }),
         ExpressionKind::Constructor(constructor) => {
             ExpressionKind::Constructor(Box::new(ConstructorExpression {
                 fields: vecmap(constructor.fields, |(name, expr)| {
@@ -733,6 +727,12 @@ fn remove_interned_in_statement_kind(
             expression: remove_interned_in_expression(interner, let_statement.expression),
             r#type: remove_interned_in_unresolved_type(interner, let_statement.r#type),
             ..let_statement
+        }),
+        StatementKind::Constrain(constrain) => StatementKind::Constrain(ConstrainStatement {
+            arguments: vecmap(constrain.arguments, |expr| {
+                remove_interned_in_expression(interner, expr)
+            }),
+            ..constrain
         }),
         StatementKind::Expression(expr) => {
             StatementKind::Expression(remove_interned_in_expression(interner, expr))
