+++ noir/noir-repo/compiler/noirc_frontend/src/hir/comptime/hir_to_display_ast.rs
@@ -8,7 +8,7 @@ use crate::ast::{
     MemberAccessExpression, MethodCallExpression, Path, PathKind, PathSegment, Pattern,
     PrefixExpression, UnresolvedType, UnresolvedTypeData, UnresolvedTypeExpression,
 };
-use crate::ast::{ConstrainExpression, Expression, Statement, StatementKind};
+use crate::ast::{ConstrainStatement, Expression, Statement, StatementKind};
 use crate::hir_def::expr::{
     HirArrayLiteral, HirBlockExpression, HirExpression, HirIdent, HirLiteral,
 };
@@ -31,6 +31,20 @@ impl HirStatement {
                 let r#type = interner.id_type(let_stmt.expression).to_display_ast();
                 let expression = let_stmt.expression.to_display_ast(interner);
                 StatementKind::new_let(pattern, r#type, expression, let_stmt.attributes.clone())
+            }
+            HirStatement::Constrain(constrain) => {
+                let expr = constrain.0.to_display_ast(interner);
+                let mut arguments = vec![expr];
+                if let Some(message) = constrain.2 {
+                    arguments.push(message.to_display_ast(interner));
+                }
+
+                // TODO: Find difference in usage between Assert & AssertEq
+                StatementKind::Constrain(ConstrainStatement {
+                    kind: ConstrainKind::Assert,
+                    arguments,
+                    span,
+                })
             }
             HirStatement::Assign(assign) => StatementKind::Assign(AssignStatement {
                 lvalue: assign.lvalue.to_display_ast(interner),
@@ -165,20 +179,6 @@ impl HirExpression {
                     }),
                     is_macro_call: false,
                 }))
-            }
-            HirExpression::Constrain(constrain) => {
-                let expr = constrain.0.to_display_ast(interner);
-                let mut arguments = vec![expr];
-                if let Some(message) = constrain.2 {
-                    arguments.push(message.to_display_ast(interner));
-                }
-
-                // TODO: Find difference in usage between Assert & AssertEq
-                ExpressionKind::Constrain(ConstrainExpression {
-                    kind: ConstrainKind::Assert,
-                    arguments,
-                    span,
-                })
             }
             HirExpression::Cast(cast) => {
                 let lhs = cast.lhs.to_display_ast(interner);
