+++ noir/noir-repo/compiler/noirc_frontend/src/parser/parser/expression.rs
@@ -3,10 +3,9 @@ use noirc_errors::Span;
 
 use crate::{
     ast::{
-        ArrayLiteral, BlockExpression, CallExpression, CastExpression, ConstrainExpression,
-        ConstrainKind, ConstructorExpression, Expression, ExpressionKind, Ident, IfExpression,
-        IndexExpression, Literal, MatchExpression, MemberAccessExpression, MethodCallExpression,
-        Statement, TypePath, UnaryOp, UnresolvedType,
+        ArrayLiteral, BlockExpression, CallExpression, CastExpression, ConstructorExpression,
+        Expression, ExpressionKind, Ident, IfExpression, IndexExpression, Literal, MatchExpression,
+        MemberAccessExpression, MethodCallExpression, Statement, TypePath, UnaryOp, UnresolvedType,
     },
     parser::{labels::ParsingRuleLabel, parser::parse_many::separated_by_comma, ParserErrorReason},
     token::{Keyword, Token, TokenKind},
@@ -654,7 +653,6 @@ impl<'a> Parser<'a> {
     ///     | ArrayExpression
     ///     | SliceExpression
     ///     | BlockExpression
-    ///     | ConstrainExpression
     ///
     /// QuoteExpression = 'quote' '{' token* '}'
     ///
@@ -696,10 +694,6 @@ impl<'a> Parser<'a> {
 
         if let Some(kind) = self.parse_block() {
             return Some(ExpressionKind::Block(kind));
-        }
-
-        if let Some(constrain) = self.parse_constrain_expression() {
-            return Some(ExpressionKind::Constrain(constrain));
         }
 
         None
@@ -806,49 +800,6 @@ impl<'a> Parser<'a> {
         }
     }
 
-    /// ConstrainExpression
-    ///     = 'constrain' Expression
-    ///     | 'assert' Arguments
-    ///     | 'assert_eq' Arguments
-    pub(super) fn parse_constrain_expression(&mut self) -> Option<ConstrainExpression> {
-        let start_span = self.current_token_span;
-        let kind = self.parse_constrain_kind()?;
-
-        Some(match kind {
-            ConstrainKind::Assert | ConstrainKind::AssertEq => {
-                let arguments = self.parse_arguments();
-                if arguments.is_none() {
-                    self.expected_token(Token::LeftParen);
-                }
-                let arguments = arguments.unwrap_or_default();
-
-                ConstrainExpression { kind, arguments, span: self.span_since(start_span) }
-            }
-            ConstrainKind::Constrain => {
-                self.push_error(ParserErrorReason::ConstrainDeprecated, self.previous_token_span);
-
-                let expression = self.parse_expression_or_error();
-                ConstrainExpression {
-                    kind,
-                    arguments: vec![expression],
-                    span: self.span_since(start_span),
-                }
-            }
-        })
-    }
-
-    fn parse_constrain_kind(&mut self) -> Option<ConstrainKind> {
-        if self.eat_keyword(Keyword::Assert) {
-            Some(ConstrainKind::Assert)
-        } else if self.eat_keyword(Keyword::AssertEq) {
-            Some(ConstrainKind::AssertEq)
-        } else if self.eat_keyword(Keyword::Constrain) {
-            Some(ConstrainKind::Constrain)
-        } else {
-            None
-        }
-    }
-
     /// Block = '{' Statement* '}'
     pub(super) fn parse_block(&mut self) -> Option<BlockExpression> {
         if !self.eat_left_brace() {
@@ -898,8 +849,8 @@ mod tests {
 
     use crate::{
         ast::{
-            ArrayLiteral, BinaryOpKind, ConstrainKind, Expression, ExpressionKind, Literal,
-            StatementKind, UnaryOp, UnresolvedTypeData,
+            ArrayLiteral, BinaryOpKind, Expression, ExpressionKind, Literal, StatementKind,
+            UnaryOp, UnresolvedTypeData,
         },
         parser::{
             parser::tests::{
@@ -1797,46 +1748,5 @@ mod tests {
             panic!("Expected unquote");
         };
         assert_eq!(expr.kind.to_string(), "((1 + 2))");
-    }
-
-    #[test]
-    fn parses_assert() {
-        let src = "assert(true, \"good\")";
-        let expression = parse_expression_no_errors(src);
-        let ExpressionKind::Constrain(constrain) = expression.kind else {
-            panic!("Expected constrain expression");
-        };
-        assert_eq!(constrain.kind, ConstrainKind::Assert);
-        assert_eq!(constrain.arguments.len(), 2);
-    }
-
-    #[test]
-    fn parses_assert_eq() {
-        let src = "assert_eq(1, 2, \"bad\")";
-        let expression = parse_expression_no_errors(src);
-        let ExpressionKind::Constrain(constrain) = expression.kind else {
-            panic!("Expected constrain expression");
-        };
-        assert_eq!(constrain.kind, ConstrainKind::AssertEq);
-        assert_eq!(constrain.arguments.len(), 3);
-    }
-
-    #[test]
-    fn parses_constrain() {
-        let src = "
-        constrain 1
-        ^^^^^^^^^
-        ";
-        let (src, span) = get_source_with_error_span(src);
-        let mut parser = Parser::for_str(&src);
-        let expression = parser.parse_expression_or_error();
-        let ExpressionKind::Constrain(constrain) = expression.kind else {
-            panic!("Expected constrain expression");
-        };
-        assert_eq!(constrain.kind, ConstrainKind::Constrain);
-        assert_eq!(constrain.arguments.len(), 1);
-
-        let reason = get_single_error_reason(&parser.errors, span);
-        assert!(matches!(reason, ParserErrorReason::ConstrainDeprecated));
     }
 }
