+++ noir/noir-repo/compiler/noirc_frontend/src/tests.rs
@@ -900,6 +900,7 @@ fn find_lambda_captures(stmts: &[StmtId], interner: &NodeInterner, result: &mut 
             HirStatement::Expression(expr_id) => expr_id,
             HirStatement::Let(let_stmt) => let_stmt.expression,
             HirStatement::Assign(assign_stmt) => assign_stmt.expression,
+            HirStatement::Constrain(constr_stmt) => constr_stmt.0,
             HirStatement::Semi(semi_expr) => semi_expr,
             HirStatement::For(for_loop) => for_loop.block,
             HirStatement::Loop(block) => block,
@@ -4053,159 +4054,6 @@ fn infers_lambda_argument_from_call_function_type_in_generic_call() {
 }
 
 #[test]
-fn infers_lambda_argument_from_call_function_type_as_alias() {
-    let src = r#"
-    struct Foo {
-        value: Field,
-    }
-
-    type MyFn = fn(Foo) -> Field;
-
-    fn call(f: MyFn) -> Field {
-        f(Foo { value: 1 })
-    }
-
-    fn main() {
-        let _ = call(|foo| foo.value);
-    }
-    "#;
-    assert_no_errors(src);
-}
-
-#[test]
-fn infers_lambda_argument_from_function_return_type() {
-    let src = r#"
-    pub struct Foo {
-        value: Field,
-    }
-
-    pub fn func() -> fn(Foo) -> Field {
-        |foo| foo.value
-    }
-
-    fn main() {
-    }
-    "#;
-    assert_no_errors(src);
-}
-
-#[test]
-fn infers_lambda_argument_from_function_return_type_multiple_statements() {
-    let src = r#"
-    pub struct Foo {
-        value: Field,
-    }
-
-    pub fn func() -> fn(Foo) -> Field {
-        let _ = 1;
-        |foo| foo.value
-    }
-
-    fn main() {
-    }
-    "#;
-    assert_no_errors(src);
-}
-
-#[test]
-fn infers_lambda_argument_from_function_return_type_when_inside_if() {
-    let src = r#"
-    pub struct Foo {
-        value: Field,
-    }
-
-    pub fn func() -> fn(Foo) -> Field {
-        if true {
-            |foo| foo.value
-        } else {
-            |foo| foo.value
-        }
-    }
-
-    fn main() {
-    }
-    "#;
-    assert_no_errors(src);
-}
-
-#[test]
-fn infers_lambda_argument_from_variable_type() {
-    let src = r#"
-    pub struct Foo {
-        value: Field,
-    }
-
-    fn main() {
-      let _: fn(Foo) -> Field = |foo| foo.value;
-    }
-    "#;
-    assert_no_errors(src);
-}
-
-#[test]
-fn infers_lambda_argument_from_variable_alias_type() {
-    let src = r#"
-    pub struct Foo {
-        value: Field,
-    }
-
-    type FooFn = fn(Foo) -> Field;
-
-    fn main() {
-      let _: FooFn = |foo| foo.value;
-    }
-    "#;
-    assert_no_errors(src);
-}
-
-#[test]
-fn infers_lambda_argument_from_variable_double_alias_type() {
-    let src = r#"
-    pub struct Foo {
-        value: Field,
-    }
-
-    type FooFn = fn(Foo) -> Field;
-    type FooFn2 = FooFn;
-
-    fn main() {
-      let _: FooFn2 = |foo| foo.value;
-    }
-    "#;
-    assert_no_errors(src);
-}
-
-#[test]
-fn infers_lambda_argument_from_variable_tuple_type() {
-    let src = r#"
-    pub struct Foo {
-        value: Field,
-    }
-
-    fn main() {
-      let _: (fn(Foo) -> Field, _) = (|foo| foo.value, 1);
-    }
-    "#;
-    assert_no_errors(src);
-}
-
-#[test]
-fn infers_lambda_argument_from_variable_tuple_type_aliased() {
-    let src = r#"
-    pub struct Foo {
-        value: Field,
-    }
-
-    type Alias = (fn(Foo) -> Field, Field);
-
-    fn main() {
-      let _: Alias = (|foo| foo.value, 1);
-    }
-    "#;
-    assert_no_errors(src);
-}
-
-#[test]
 fn regression_7088() {
     // A test for code that initially broke when implementing inferring
     // lambda parameter types from the function type related to the call
@@ -4345,22 +4193,4 @@ fn call_function_alias_type() {
     }
     "#;
     assert_no_errors(src);
-}
-
-#[test]
-fn errors_on_if_without_else_type_mismatch() {
-    let src = r#"
-    fn main() {
-        if true { 
-            1 
-        }
-    }
-    "#;
-    let errors = get_program_errors(src);
-    assert_eq!(errors.len(), 1);
-
-    let CompilationError::TypeError(TypeCheckError::Context { err, .. }) = &errors[0].0 else {
-        panic!("Expected a Context error");
-    };
-    assert!(matches!(**err, TypeCheckError::TypeMismatch { .. }));
 }
