+++ noir/noir-repo/test_programs/execution_report.sh
@@ -8,7 +8,7 @@ base_path="$current_dir/execution_success"
 # Tests to be profiled for execution report
 tests_to_profile=("sha256_regression" "regression_4709" "ram_blowup_regression" "global_var_regression_entry_points")
 
-echo "[" > $current_dir/execution_report.json
+echo "{\"execution_reports\": [ " > $current_dir/execution_report.json
 
 # If there is an argument that means we want to generate a report for only the current directory
 if [ "$1" == "1" ]; then
@@ -70,7 +70,7 @@ for dir in ${tests_to_profile[@]}; do
                 printf "%.3f\n", 0
         }' <<<"${TIMES[@]}")
 
-    jq -rc "{name: \"$PACKAGE_NAME\", value: \""$AVG_TIME"\" | tonumber, unit: \"s\"}" --null-input >> $current_dir/execution_report.json
+    jq -rc "{artifact_name: \"$PACKAGE_NAME\", time: \""$AVG_TIME"s\"}" --null-input >> $current_dir/execution_report.json
 
     if (($ITER != $NUM_ARTIFACTS)); then
         echo "," >> $current_dir/execution_report.json
@@ -81,4 +81,4 @@ for dir in ${tests_to_profile[@]}; do
     ITER=$(( $ITER + 1 ))
 done
 
-echo "]" >> $current_dir/execution_report.json
+echo "]}" >> $current_dir/execution_report.json
