--- noir/noir-repo/test_programs/execution_success/signed_inactive_division_by_zero/Nargo.toml
@@ -1,6 +1,0 @@
-[package]
-name = "signed_inactive_division_by_zero"
-type = "bin"
-authors = [""]
-
-[dependencies]
