--- noir/noir-repo/test_programs/execution_success/signed_inactive_division_by_zero/Prover.toml
@@ -1,3 +1,0 @@
-a = "1"
-b = "0"
-condition = false
