--- noir/noir-repo/test_programs/execution_success/signed_inactive_division_by_zero/src/main.nr
@@ -1,8 +1,0 @@
-fn main(a: i32, b: i32, condition: bool) -> pub i32 {
-    if condition {
-        // If `condition` is not set then we should not trigger an assertion failure here.
-        a / b
-    } else {
-        0
-    }
-}
